historical/requirements.txt
shopifyapi
pandas
pandas-gbq
google-cloud-bigquery
google-auth
google-cloud-core
python-dotenv
